Move OCR option to own TextExtractor class
